requirements-test.txt
flake8==2.2.3
pytest==2.6.1
pytest-cov==1.8.0
pytest-django==2.6.2
